+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 6.1.0 under development
 
 - Enh #79: Add debug collector for yiisoft/yii-debug (@xepozz)
+- Bug #82: Fixed find for layout file due to compatibility with `yiisoft/view` (@rustamwin)
 
 ## 6.0.0 February 16, 2023